Moved code back into __init__.py
