--- frontend/dashboard/gui.py
@@ -1,1 +1,0 @@
-from widgets import *
