--- frontend/frontend.py
@@ -1,6 +1,0 @@
-class Frontend:
-    dashboard = Dashboard()
-    viewfinder = ViewFinder()
-    while True:
-        dashboard.update()
-        viewfinder.update()
