+++ frontend/viewfinder/__init__.py
@@ -1,1 +1,14 @@
-from .viewfinder import Viewfinder
+# kwu 11/16/2019
+
+class Viewfinder:
+    # constructor
+    def __init__(self):
+        print("Initializing Viewfinder")
+
+    # find the camera
+    def find_camera(self):
+        print("testtt")
+
+    # updates image shown
+    def update(self):
+        print("should updataete")
