--- frontend/viewfinder/viewfinder.py
@@ -1,11 +1,0 @@
-#kwu 11/16/2019
-class Viewfinder:
-    #constructor
-    def __init__(self):
-        print("Initializing Viewfinder")
-    #find the camera
-    def findCamera(self):
-        print("testtt")
-    #updates image shown
-    def update(self):
-        print("should updataete")
