--- main.py
@@ -1,7 +1,0 @@
-from tkinter import *
-
-root = Tk()
-
-root.geometry(str(root.winfo_screenwidth())+'x'+str(root.winfo_screenheight()))
-
-root.mainloop()
